+++ Classes/TripPurposeDelegate.h
@@ -88,7 +88,7 @@
 @optional
 - (void)didCancelPurpose;
 - (void)didCancelNote;
-- (void)didPickPurpose:(unsigned int)index;
+- (void)didPickPurpose:(NSInteger)index;
 - (void)didPickNoteType:(NSNumber *)index;
 - (void)didEnterNoteDetails:(NSString *)details;
 - (void)didEnterTripDetails:(NSString *)details;
